Clean repo, enable -O1, refactor host, add errata & local test harness
MCU (mcu_part/v1.2.X):
- Enable -O1 optimization (was -O0/disabled): 1677 -> 1570 program words (-6.4%),
  behaviour unchanged. Updated nbproject/configurations.xml.
- Document the 32-word flash-row alignment requirement directly in FLASH_Write()
  (the "address does not roll over" behaviour is normal PIC16F1 row-write, not a
  silicon erratum).
- Add build.sh for XC8 CLI size checks outside MPLAB X.

Host (host_part/cpp):
- Refactor duplicated send/recv/check pattern into Transact(),
  TransactExpectFull(), MakeReadRequest() helpers; wire into read paths.
- Add mock_mcu.py (protocol mock + self-test) and test_e2e.sh (socat-based
  end-to-end: detect + flash led.hex + full readback, all passing).

Docs:
- Add PIC16(L)F1788/1789 silicon errata DS80000575.

Cleanup:
- Remove tracked build artifacts and scratch dumps (117 -> 62 tracked files).
- Rewrite .gitignore to cover MPLAB X and host build output.

+++ mcu_part/v1.2.X/internal_flash.c
@@ -79,7 +79,19 @@ bool FLASH_Write(uint8_t *buffer){
 
     INTCONbits.GIE = 0; // Disable interrupts
 
-    writeAddr = (((buffer[2] << 8) & 0xFF00) | (buffer[3] & 0x00FF));
+    // NOTE (flash row behaviour / alignment requirement):
+    // On PIC16F1 the program-memory write engine latches a whole row of
+    // 32 words. The row is selected by EEADRH:EEADRL[13:5]; only the low
+    // 5 bits EEADRL[4:0] pick one of the 32 write latches inside that row.
+    // During the latch-loading loop below the address is incremented, but
+    // if a block does NOT start on a 32-word boundary (0x20) the low bits
+    // wrap around inside the SAME row instead of advancing to the next one
+    // (looks like the address "does not roll over" / "jumps to +1").
+    // Workaround: the host always sends blocks of 32 words aligned to 0x20
+    // (see fw_converter.cpp and docs/BL_protocol.txt "aligned to x32").
+    // The XC8 memory model reserves the BL area accordingly:
+    //   default,-0-37FF,-3FE0-3FFF
+    writeAddr = (uint16_t)(((buffer[2] << 8) & 0xFF00) | (buffer[3] & 0x00FF));
         EEADRH = (uint8_t)((writeAddr & 0xFF00)>>8);
         EEADRL = (uint8_t)(writeAddr & 0x00FF);
